backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
kafka-python==2.0.2
scikit-learn==1.4.2
numpy==1.26.4
torch==2.3.0
torch-geometric==2.5.3
boto3==1.34.100
slack-sdk==3.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
websockets==12.0
redis==5.0.4
prometheus-client==0.20.0
httpx==0.27.0
python-dotenv==1.0.1
pandas==2.2.2
joblib==1.4.0